requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
openai==1.14.3
python-dotenv==1.0.1